+++ catgraph/src/frobenius/mod.rs
@@ -9,6 +9,7 @@ mod trait_impl;
 
 pub use morphism_system::{Contains, InterpretableMorphism, MorphismSystem};
 pub use operations::{
-    FrobeniusMorphism, FrobeniusOperation, from_decomposition, special_frobenius_morphism,
+    FrobeniusMorphism, FrobeniusOperation, frobenius_to_cospan, from_decomposition,
+    special_frobenius_morphism,
 };
 pub use trait_impl::Frobenius;